refactor: 요금정책 계산 static 메소드 분리
