+++ src/main/java/nextstep/subway/path/application/PathFinder.java
@@ -4,7 +4,8 @@ import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.line.domain.Sections;
+import nextstep.subway.path.application.farepolicy.FarePolicy;
+import nextstep.subway.path.application.farepolicy.FarePolicyElements;
 import nextstep.subway.path.dto.PathVertexStation;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
